# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/auth/check-logs-auth": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Используется Caddy как точка forward_auth перед проксированием на Dozzle.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Проверка доступа к логам (Caddy forward_auth)",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "missing or invalid token"
                    },
                    "403": {
                        "description": "insufficient permissions (not admin)"
                    }
                }
            }
        },
...
        "/auth/google": {
            "post": {
                "description": "Authenticates a user using Google OAuth ID Token. Returns an access token and sets a HTTP-only refresh token cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth Login",
                "parameters": [
                    {
                        "description": "Google ID token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.GoogleAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "invalid JSON",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "google auth failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "account is blocked",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Clears the refresh token cookie and revokes the active session",
                "tags": [
                    "auth"
                ],
                "summary": "Log out",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Reissues access token and refresh token using refresh_token cookie",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.TokenResponse"
                        }
                    },
                    "401": {
                        "description": "refresh token missing or invalid session",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Создаёт нового пользователя с минимальной ролью (pharmacist).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя по email",
                "parameters": [
                    {
                        "description": "Email для регистрации",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "email is required",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "email already registered",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "failed to register",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/send-code": {
            "post": {
                "description": "Sends a 6-digit OTP code to the user's email for passwordless login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Send OTP code",
                "parameters": [
                    {
                        "description": "Email address to send OTP code to",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.SendCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.SendCodeResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body or email is required",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "user with this email not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "too many requests, try again later",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "failed to send code",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify-code": {
            "post": {
                "description": "Verifies the OTP code sent to the user's email. Returns an access token and sets a HTTP-only refresh token cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify OTP code",
                "parameters": [
                    {
                        "description": "Email and OTP code to verify",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.VerifyCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "invalid request body or missing fields",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "invalid or expired code",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "failed to verify code",
                        "schema": {
                            "$ref": "#/definitions/github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_ima_diplom-backend_internal_handler_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_ima_diplom-backend_internal_handler_dto.GoogleAuthRequest": {
            "type": "object",
            "properties": {
                "id_token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ima_diplom-backend_internal_handler_dto.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "github_com_ima_diplom-backend_internal_handler_dto.RegisterResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_ima_diplom-backend_internal_handler_dto.SendCodeRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "github_com_ima_diplom-backend_internal_handler_dto.SendCodeResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_ima_diplom-backend_internal_handler_dto.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_ima_diplom-backend_internal_handler_dto.VerifyCodeRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
...
    }
}